# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Whispick API — anonymous video recommendation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whisps:
    get:
      operationId: listWhisps
      tags: [whisps]
      summary: List all whisps sent by the current user
      parameters:
        - in: query
          name: status
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of whisps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Whisp"
    post:
      operationId: createWhisp
      tags: [whisps]
      summary: Send a new whisp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhispInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Whisp"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /whisps/{id}:
    get:
      operationId: getWhisp
      tags: [whisps]
      summary: Get a single whisp with tracking timeline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Whisp detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhispDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteWhisp
      tags: [whisps]
      summary: Delete a whisp
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /admin/whisps:
    get:
      operationId: adminListWhisps
      tags: [admin]
      summary: List all whisps across all users (admin only)
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: deliveryMethod
          schema:
            type: string
          required: false
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: pageSize
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Whisps page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWhispListResponse"
  /admin/whisps/{id}:
    get:
      operationId: adminGetWhisp
      tags: [admin]
      summary: Get full detail on any whisp (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Whisp detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWhispDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: adminDeleteWhisp
      tags: [admin]
      summary: Remove a whisp (moderation) (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /whisper-groups:
    get:
      operationId: listWhisperGroups
      tags: [whisperGroups]
      summary: List the current user's saved whisper groups
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhisperGroupSummary"
    post:
      operationId: createWhisperGroup
      tags: [whisperGroups]
      summary: Create a new whisper group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhisperGroupNameInput"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhisperGroupSummary"
...
  /whisper-groups/sends/{groupSendId}:
    get:
      operationId: getGroupWhispSend
      tags: [whisperGroups]
      summary: Per-member breakdown of one past group whisper send
      parameters:
        - in: path
          name: groupSendId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group send detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupWhispSendDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /whisper-groups/{id}:
    get:
      operationId: getWhisperGroup
      tags: [whisperGroups]
      summary: Get a whisper group and its members
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhisperGroupDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: renameWhisperGroup
      tags: [whisperGroups]
      summary: Rename a whisper group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhisperGroupNameInput"
      responses:
        "200":
          description: Renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhisperGroup"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteWhisperGroup
      tags: [whisperGroups]
      summary: Delete a whisper group and its saved member list
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /whisper-groups/{id}/members:
    post:
      operationId: addWhisperGroupMembers
      tags: [whisperGroups]
      summary: Add one or more members to a whisper group (manual entry or a Contact Picker batch)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWhisperGroupMembersInput"
      responses:
        "201":
          description: Updated member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhisperGroupMember"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /whisper-groups/{id}/members/{memberId}:
    delete:
      operationId: removeWhisperGroupMember
      tags: [whisperGroups]
      summary: Remove a member from a whisper group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: memberId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /whisper-groups/{id}/send:
    post:
      operationId: sendGroupWhisp
      tags: [whisperGroups]
      summary: Send a whisp to every deliverable member of a whisper group, anonymously
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendGroupWhispInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendGroupWhispResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "402":
          description: Payment required — Whisper Link plan limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # POST /media/upload (multipart/form-data video+thumbnail+durationSeconds
  # → UploadedVideo | 400 | 503 ApiError) is intentionally not modeled here:
  # a `format: binary` request body generates `File`/`Blob`-typed Zod schemas
  # that don't compile in lib/api-zod's Node-only (no DOM lib) project, and
  # this endpoint is called via a hand-written fetch from the frontend
  # anyway (multipart isn't a great fit for the generated JSON client) — see
  # artifacts/whispick/src/lib/uploadMedia.ts.
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
    Whisp:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        videoUrl:
          type: string
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoEmbedUrl:
          type: ["string", "null"]
        videoStartSeconds:
          type: ["integer", "null"]
        videoPlatform:
          type: ["string", "null"]
        videoTranscript:
          type: ["string", "null"]
        uploadedVideoId:
          type: ["string", "null"]
        deliveryMethod:
          type: string
        whisperChannel:
          type: ["string", "null"]
        circleId:
          type: ["string", "null"]
        recipientEmail:
          type: ["string", "null"]
        recipientPhone:
          type: ["string", "null"]
        anonymousNote:
          type: ["string", "null"]
        senderAlias:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        status:
          type: string
        publicToken:
          type: string
        scheduledAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        openedAt:
          type: ["string", "null"]
        watchedAt:
          type: ["string", "null"]
        revealRequested:
          type: boolean
        revealAccepted:
          type: ["boolean", "null"]
        appreciationResponse:
          type: ["string", "null"]
        appreciationRespondedAt:
          type: ["string", "null"]
        aiTakeaway:
          type: ["string", "null"]
        aiTakeawayStatus:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - senderId
        - videoUrl
        - deliveryMethod
        - status
        - publicToken
        - revealRequested
        - createdAt
    WhispInput:
      type: object
      properties:
        videoUrl:
          type: ["string", "null"]
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoEmbedUrl:
          type: ["string", "null"]
        videoStartSeconds:
          type: ["integer", "null"]
        videoPlatform:
          type: ["string", "null"]
        uploadedVideoId:
          type: ["string", "null"]
          description: An id from the sender's Media Library — an alternative to videoUrl. One of the two is required.
        deliveryMethod:
          type: string
        whisperChannel:
          type: ["string", "null"]
        circleId:
          type: ["string", "null"]
        recipientEmail:
          type: ["string", "null"]
        recipientPhone:
          type: ["string", "null"]
        anonymousNote:
          type: ["string", "null"]
        senderAlias:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
      required:
        - deliveryMethod
    WhispDetail:
      type: object
      properties:
        whisp:
          $ref: "#/components/schemas/Whisp"
        trackingEvents:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/WhispReply"
      required:
        - whisp
        - trackingEvents
        - replies
...
    AdminWhispListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminWhispListItem"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - items
        - total
        - page
        - pageSize
    AdminWhispDetail:
      type: object
      properties:
        whisp:
          $ref: "#/components/schemas/Whisp"
        senderId:
          type: ["string", "null"]
        senderEmail:
          type: ["string", "null"]
        senderFullName:
          type: ["string", "null"]
        trackingEvents:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/WhispReply"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/WhispCategory"
      required:
        - whisp
        - trackingEvents
        - replies
        - categories
...
    WhisperGroup:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - ownerId
        - name
        - createdAt
    WhisperGroupSummary:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        createdAt:
          type: string
        memberCount:
          type: integer
      required:
        - id
        - ownerId
        - name
        - createdAt
        - memberCount
...
    WhisperGroupDetail:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/WhisperGroupMember"
      required:
        - id
        - ownerId
        - name
        - createdAt
        - members
    WhisperGroupNameInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...
    AddWhisperGroupMembersInput:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMemberInput"
      required:
        - members
...
    GroupWhispSendDetail:
      type: object
      properties:
        groupSendId:
          type: string
        groupName:
          type: ["string", "null"]
        video:
          $ref: "#/components/schemas/GroupWhispSendVideo"
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupWhispSendMember"
      required:
        - groupSendId
        - video
        - members
...